서버 업데이트, 클라이언트 업데이트
1. 유저 데드레커닝 구현 - 1초에 10번 패킷을 보내 위치와 방향을 업데이트한다.
2. 클라이언트 - 서버로 부터 TargetPosition을 받아 벡터 이동을 한다.

앞으로 할일
1.유저간 동기화가 잘되면 위치 업데이트 주기를 길게 잡아도 된다. (Test를 통한 시간 정하기)
2. 더미 클라이언트 기능 객체화
3. 3D 클라이언트와 연동
4. Object간 충돌체크

+++ server/Client/Win32Project1/Object.h
@@ -54,7 +54,7 @@ public:
 		if (objectStatus.hp < 0) is_using = false; 
 	}
 	void minusHp(const WORD hp) { objectStatus.hp -= hp; }
-	const WORD getHp()const { return objectStatus.hp; }
+	const int getHp()const { return objectStatus.hp; }
 
 
 	ObjectStatus& getStatus(){ return objectStatus; }
